requirements.txt
# HTTP 요청 라이브러리
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 이미지 처리 라이브러리
Pillow>=10.0.0

# 웹 애플리케이션 프레임워크
streamlit>=1.28.0

# OpenAI API 클라이언트
openai>=1.35.0

# 추가 유용한 패키지들
pandas>=2.0.0
numpy>=1.24.0 